+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 
 ## (UNRELEASED)
 
+### FEATURES:
+* provisioner/ChefClient: add support for packer's chef client provisioner
+* provisioner/ChefSolo: add support for packer's chef solo provisioner
+* provisioner/Converge: add support for packer's converge provisioner
+* provisioner/PowerShell: add support for packer's powershell provisioner
+* provisioner/PuppetMasterless: add support for packer's puppet masterless provisioner
+* provisioner/PuppetServer: add support for packer's puppet server provisioner
+* provisioner/WindowsShell: add support for packer's windows shell provisioner
+* provisioner/WindowsRestart: add support for packer's windows restart provisioner
 
 ## 0.2.2 (August 13, 2017)